# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/matches": {
            "get": {
                "tags": [
                    "Match"
                ],
                "summary": "Get a listing matches with details based on season year and week number.",
                "description": "Get all Matches",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "year",
                        "in": "query",
                        "description": "season year",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "week_num",
                        "in": "query",
                        "description": "week number",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Match"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/seasonMatches": {
            "get": {
                "tags": [
                    "Match"
                ],
                "summary": "Get a listing of the Matches grouped by season.",
                "description": "Get all Matches",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Match"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/matches/{id}": {
            "get": {
                "tags": [
                    "Match"
                ],
                "summary": "Display the specified Match with its details by ID.",
                "description": "Get Match",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Match",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Match"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Match": {
            "required": [
                "title",
                "description",
                "title_ar",
                "description_ar",
                "week_id"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "title": {
                    "description": "title",
                    "type": "string"
                },
                "description": {
                    "description": "description",
                    "type": "string"
                },
                "title_ar": {
                    "description": "title_ar",
                    "type": "string"
                },
                "description_ar": {
                    "description": "description_ar",
                    "type": "string"
                },
                "image": {
                    "description": "image",
                    "type": "string"
                },
                "video": {
                    "description": "video",
                    "type": "string"
                },
                "week_num": {
                    "description": "week_num",
                    "type": "integer",
                    "format": "int32"
                },
                "year": {
                    "description": "year",
                    "type": "string"
                }
            }
        }
    }
}